refactor: updated to use new code layout (#2)
