+++ azure/plugin.go
@@ -1,31 +1,20 @@
 package azure
 
 import (
-	"github.com/turbot/tailpipe-plugin-azure/azure_table"
-	"github.com/turbot/tailpipe-plugin-sdk/table"
-	"log/slog"
-	"time"
-
-	"github.com/turbot/tailpipe-plugin-azure/azure_source"
+	"github.com/turbot/tailpipe-plugin-azure/sources"
+	"github.com/turbot/tailpipe-plugin-azure/tables"
 	"github.com/turbot/tailpipe-plugin-sdk/plugin"
 	"github.com/turbot/tailpipe-plugin-sdk/row_source"
+	"github.com/turbot/tailpipe-plugin-sdk/table"
 )
 
-type Plugin struct {
-	plugin.PluginBase
-}
-
 func NewPlugin() (plugin.TailpipePlugin, error) {
-	p := &Plugin{}
-
-	slog.Info("Azure Plugin starting")
-	time.Sleep(10 * time.Second)
-	slog.Info("Running...")
+	p := plugin.NewPlugin("azure")
 
 	err := p.RegisterResources(
 		&plugin.ResourceFunctions{
-			Tables:  []func() table.Table{azure_table.NewActivityLogTable},
-			Sources: []func() row_source.RowSource{azure_source.NewActivityLogAPISource},
+			Tables:  []func() table.Table{tables.NewActivityLogTable},
+			Sources: []func() row_source.RowSource{sources.NewActivityLogAPISource},
 		})
 
 	if err != nil {
@@ -33,8 +22,4 @@ func NewPlugin() (plugin.TailpipePlugin, error) {
 	}
 
 	return p, nil
-}
-
-func (t *Plugin) Identifier() string {
-	return "azure"
 }
