+++ sources/activity_log_api_source.go
@@ -1,4 +1,4 @@
-package azure_source
+package sources
 
 import (
 	"context"
