+++ tables/activity_log_table_config.go
@@ -1,4 +1,4 @@
-package azure_table
+package tables
 
 type ActivityLogTableConfig struct{}
